pyproject.toml
[project]
name = "investallies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-ckeditor>=6.7.3",
    "django-import-export>=4.3.14",
    "django-modeltranslation>=0.19.17",
    "faker>=38.2.0",
    "pillow>=12.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.8",
]
